+++ DALTON/include/inftra.h
@@ -1,10 +1,12 @@
 C --- File: inftra.h ---
 C     Global information for the sirius/sirtra.F 2-el integral transforamtion modules
 C
+      REAL*8  THRP,   THRQ
+      INTEGER LSRTAO, IPRTRA, LBUF, NIBUF, NBITS, NCDTRA, FCKTRA_TYPE
       LOGICAL USEDRC, NEWTRA, NEWTRA_USEDRC, USE_INTSORT
-      COMMON /INFTRA/ THRP,THRQ,
-     &                LSRTAO,IPRTRA, LBUF,NIBUF,NBITS,
-     &                USEDRC, NEWTRA,NEWTRA_USEDRC, USE_INTSORT
+      COMMON /INFTRA/ THRP,   THRQ,
+     &        LSRTAO, IPRTRA, LBUF, NIBUF, NBITS, NCDTRA, FCKTRA_TYPE,
+     &        USEDRC, NEWTRA, NEWTRA_USEDRC, USE_INTSORT
 C
 C     Name of file with 2-el integrals in AO basis.
 C     Defaults to "AOTWOINT", but can be set to e.g. "AOSR2INT" to transform